create secrets.go, move ingresSpec creation to generator.go
